bank managment system

+++ Introduction-of-c++/Module-4/Monkey.cpp
@@ -6,7 +6,7 @@ int main()
     char s[100005];
     while(cin.getline(s, 100005)){
         sort(s, s+strlen(s));
-        int len = strlen(s);
+        long long int len = strlen(s);
         for(int i = 0; i <len; i++){
             if(s[i]!=' '){
                 cout<<s[i];
